# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test API with swagger",
    "version": "1.0.0",
    "description": ""
  },
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/subject/create": {
      "post": {
        "tags": [
          "Subject"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "description": "time과 days는 배열, days는 0~6까지 일요일부터 토요일을 의미",
            "schema": {
              "$ref": "#/definitions/createSubject"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "성공적으로 해당 subject를 개설했을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "code": {
                  "type": "string",
                  "example": "519hi32hkjifb12"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/subject/join": {
      "put": {
        "tags": [
          "Subject"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "description": "참여코드 입력",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": ""
                }
              },
              "required": [
                "code"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공시 success, subject 객체 반환",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "subject": {
                  "$ref": "#/definitions/subject "
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "404": {
            "description": "해당하는 코드의 강의가 없는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "codeValidation": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "409": {
            "description": "이미 참여한 수업일 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "subjectExist": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/subject/info/{id}": {
      "get": {
        "tags": [
          "Subject"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 해당 강의 정보 받아온 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "subject": {
                  "$ref": "#/definitions/subject"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "404": {
            "description": "해당 subject가 존재하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "existSubject": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/subject/info/update/{id}": {
      "put": {
        "tags": [
          "Subject"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "any"
                },
                "start_period": {
                  "type": "string",
                  "example": "any"
                },
                "end_period": {
                  "type": "string",
                  "example": "any"
                },
                "start_time": {
                  "type": "string",
                  "example": "any"
                },
                "end_time": {
                  "type": "string",
                  "example": "any"
                },
                "days": {
                  "type": "string",
                  "example": "any"
                },
                "introURL": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 해당 강의 정보를 수정했을 경우",
            "schema": {
              "$ref": "#/definitions/subject"
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우,  \n해당 과목을 만든 본인이 아닐 경우, success: false, isValidProfessor: false 반환",
            "schema": {
              "type": "object",
              "properties": {
                "교수아님": {
                  "$ref": "#/definitions/proAuthFailed"
                },
                "본인아님": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "isValidProfessor": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "해당 강의가 존재하지 않을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "existSubject": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/subject/get/mySubjects": {
      "get": {
        "tags": [
          "Subject"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "성공 시, 본인이 수강하고 있는 강의들을 subjects로 반환",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "subjects": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/subject "
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/subject/get/upcoming": {
      "get": {
        "tags": [
          "Subject"
        ],
        "description": "현재 수업 중인 강의, 또는 곧 있을 강의 \n 얼마나 지났는 지, 또는 해당 강의까지 남은 시간 반환",
        "parameters": [],
        "responses": {
          "200": {
            "description": "현재 진행중인 수업이 있을 경우  \n오늘 내에 강의가 있을 경우  \n오늘 내에 강의가 없을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "진행중": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "inProgress": {
                      "type": "boolean",
                      "example": true
                    },
                    "subject": {
                      "type": "number",
                      "example": 0
                    },
                    "diffH": {
                      "type": "number",
                      "example": 0
                    },
                    "diffM": {
                      "type": "number",
                      "example": -15
                    }
                  }
                },
                "곧있음": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "upcoming": {
                      "type": "boolean",
                      "example": true
                    },
                    "subject": {
                      "type": "number",
                      "example": 0
                    },
                    "diffH": {
                      "type": "number",
                      "example": 0
                    },
                    "diffM": {
                      "type": "number",
                      "example": 30
                    }
                  }
                },
                "없음": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createSubject": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "캡스톤디자인"
        },
        "start_period": {
          "type": "string",
          "example": "2021-03-02"
        },
        "end_period": {
          "type": "string",
          "example": "2021-06-30"
        },
        "start_time": {
          "type": "array",
          "example": [
            "16:30",
            "18:00",
            "19:30"
          ],
          "items": {
            "type": "string"
          }
        },
        "end_time": {
          "type": "array",
          "example": [
            "18:00",
            "19:30",
            "21:00"
          ],
          "items": {
            "type": "string"
          }
        },
        "days": {
          "type": "array",
          "example": [
            1,
            1,
            1
          ],
          "items": {
            "type": "number"
          }
        }
      },
      "required": [
        "name",
        "start_period",
        "end_period"
      ]
    },
...
    "authFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isLogined": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "proAuthFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isProfessor": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "subject": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "number",
          "example": 0
        },
        "name": {
          "type": "string",
          "example": "캡스톤디자인"
        },
        "start_period": {
          "type": "string",
          "example": "2021-03-02"
        },
        "end_period": {
          "type": "string",
          "example": "2021-06-30"
        },
        "start_time": {
          "type": "array",
          "example": [
            "16:30",
            "18:00",
            "19:30"
          ],
          "items": {
            "type": "string"
          }
        },
        "end_time": {
          "type": "array",
          "example": [
            "18:00",
            "19:30",
            "21:00"
          ],
          "items": {
            "type": "string"
          }
        },
        "days": {
          "type": "array",
          "example": [
            1,
            1,
            1
          ],
          "items": {
            "type": "number"
          }
        },
        "code": {
          "type": "string",
          "example": "519hi32hkjifb12"
        },
        "lectures": {
          "type": "array",
          "example": [],
          "items": {}
        },
        "students": {
          "type": "array",
          "example": [
            0
          ],
          "items": {
            "type": "number"
          }
        },
        "introURL": {
          "type": "string",
          "example": ""
        }
      }
    },
...
  }
}